refactor(main): split asset routes and analytics middleware out of main
Move favicon, /pygments.css, and static file routes into
routers/assets.py, and the page-view tracking middleware, bot UA
detection, and track_page_view into services/analytics_middleware.py.
main.py is now a pure app factory and wiring module. No behavior,
route, header, or middleware-ordering changes.

Update moved-symbol imports in tests.

Refs #111
